modify
git-svn-id: https://svn.baidu.com/app/search/sep/trunk/anti-spam/melt@11057 fa64baa9-71d1-4fed-97ae-c15534abce97

+++ include/Run/MeltArguments.h
@@ -23,7 +23,7 @@ namespace gezi {
 		string commandInput = ""; //ci|
 		string classifierName = "LinearSvm";//cl|
 		string classifierSettings = ""; //cls|
-		int num = 0;
+		int64 num = 0;
 		int numFolds = 5;
 		int numRuns = 1;
 		int numThreads = 0;
